# powertrader-ts/backend/src/api/server.ts
import express from 'express';
import cors from 'cors';
import bodyParser from 'body-parser';
import http from 'http';
import { ConfigManager } from '../config/ConfigManager';
import { AnalyticsManager } from '../analytics/AnalyticsManager';
import { WebSocketManager } from './websocket';
import { CointradeAdapter } from '../modules/cointrade/CointradeAdapter';
import { KuCoinConnector } from '../exchanges/KuCoinConnector';
import { StrategyFactory } from '../engine/strategy/StrategyFactory';
import { BacktestEngine } from '../engine/backtest/BacktestEngine';
import { HyperOpt } from '../extensions/hyperopt/HyperOpt';
import { DeepThinker } from '../thinker/DeepThinker';
import { HistoricalData } from '../engine/backtest/HistoricalData';
import { UniswapConnector } from '../exchanges/UniswapConnector';
import { LiquidityManager } from '../defi/LiquidityManager';
import { ArbitrageScanner } from '../engine/ArbitrageScanner';
import { CorrelationMatrix } from '../engine/risk/CorrelationMatrix';
import { SentimentScanner } from '../engine/risk/SentimentScanner';
import { WhaleWatcher } from '../engine/risk/WhaleWatcher';
import swaggerUi from 'swagger-ui-express';
import fs from 'fs';
import path from 'path';

const app = express();
const port = 3000;

app.use(cors());
app.use(bodyParser.json());

const config = ConfigManager.getInstance();

// OpenAPI Docs
try {
    const swaggerDoc = JSON.parse(fs.readFileSync(path.join(__dirname, '../../swagger.json'), 'utf-8'));
    app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDoc));
} catch (e) {
    console.warn("Could not load swagger.json");
}

// Simple Auth Middleware
const requireAuth = (req: express.Request, res: express.Response, next: express.NextFunction) => {
    const password = config.get("system.ui_password");
    if (!password) return next(); // No password set = disabled

    const authHeader = req.headers['authorization'];
    if (!authHeader || authHeader !== `Bearer ${password}`) {
        return res.status(401).json({ error: "Unauthorized" });
    }
    next();
};

app.post('/api/auth/login', (req, res) => {
    const { password } = req.body;
    const sysPwd = config.get("system.ui_password");
    if (!sysPwd || password === sysPwd) {
        res.json({ success: true, token: password });
    } else {
        res.status(401).json({ success: false });
    }
});

// Apply auth to sensitive routes (omitted from public getters if needed, but for now apply broadly)
// app.use('/api/settings', requireAuth);
// app.use('/api/system/mode', requireAuth);
const analytics = new AnalyticsManager();
const kucoin = new KuCoinConnector();

// --- API ROUTES ---

app.get('/api/dashboard', (req, res) => {
    const perf = analytics.getPerformance();
    res.json({
        account: {
            total: 10000,
            pnl: perf.pnl
        },
        trades: [
            { symbol: 'BTC', pnl: 1.2, stage: 0 },
            { symbol: 'ETH', pnl: -0.5, stage: 1 }
        ],
        execution_mode: config.get('trading.execution_mode') || 'paper'
    });
});

app.post('/api/system/mode', (req, res) => {
    const { mode } = req.body;
    if (mode === 'live' || mode === 'paper') {
        // In real app, persist this
        // config.set('trading.execution_mode', mode);
        console.log(`[API] Switched execution mode to ${mode}`);
        res.json({ success: true, mode });
    } else {
        res.status(400).json({ error: "Invalid mode" });
    }
});

app.get('/api/settings', (req, res) => {
    res.json(config.get('trading'));
});

app.post('/api/settings', (req, res) => {
    // In a real implementation: config.set('trading', req.body);
    res.json({ success: true });
});

app.get('/api/volume/:coin', (req, res) => {
    res.json({
        profile: { average: 5000, median: 4500, p90: 8000 },
        recent: [
            { timestamp: Date.now(), volume: 4200, ratio: 1.1, trend: 'increasing' }
        ]
    });
});

// Strategy Management
app.get('/api/strategies', (req, res) => {
    const strategies = StrategyFactory.getAll().map(s => ({
        name: s.name,
        interval: s.interval
    }));
    const active = config.get('trading.active_strategy') || "SMAStrategy";
    res.json({ strategies, active });
});

app.post('/api/strategies/config', (req, res) => {
    const { strategy } = req.body;
    // In real app, we would ConfigManager.set('trading.active_strategy', strategy)
    console.log(`[API] Switching active strategy to: ${strategy}`);
    res.json({ success: true, active: strategy });
});

app.post('/api/hyperopt/run', async (req, res) => {
    try {
        console.log("[API] Running HyperOpt...", req.body);
        const config = req.body;

        // Defaults
        const hyperOptConfig = {
            strategyName: config.strategy || "SMAStrategy",
            pair: `${config.symbol || "BTC"}-USD`,
            timeframe: config.timeframe || "1h",
            startDate: config.startDate ? new Date(config.startDate).getTime() : (Date.now() - (30 * 24 * 60 * 60 * 1000)),
            endDate: config.endDate ? new Date(config.endDate).getTime() : Date.now(),
            populationSize: config.populationSize || 20,
            generations: config.generations || 10,
            parameterSpace: config.parameterSpace || {
                period: { min: 5, max: 100, step: 1 }
            }
        };

        const optimizer = new HyperOpt();
        const result = await optimizer.optimize(hyperOptConfig);

        res.json(result);
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message || "HyperOpt failed" });
    }
});

// --- Advanced DeFi (Liquidity) ---
const defiConfig = config.get("defi") || {};
// We instantiate a specific Uniswap connector for DeFi operations even if active_exchange is different
const uniswap = new UniswapConnector(defiConfig.rpc_url, defiConfig.private_key);
const liquidityManager = new LiquidityManager(uniswap);

app.get('/api/defi/positions', async (req, res) => {
    try {
        const positions = await uniswap.fetchPositions();
        res.json({ positions });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

app.post('/api/defi/liquidity/compound', async (req, res) => {
    try {
        const { tokenId } = req.body;
        // Re-use existing connector instance if possible, or create new.
        // Ideally, we should have a Singleton 'LiquidityManager' service.
        // For MVP, recreating.
        const connector = new UniswapConnector(config.get("defi.rpc_url"), config.get("defi.private_key"));
        const manager = new LiquidityManager(connector);

        const result = await manager.autoCompound(Number(tokenId));
        res.json(result);
    } catch (e: any) {
        res.status(500).json({ success: false, error: e.message });
    }
});

app.post('/api/defi/liquidity/add', async (req, res) => {
    try {
        const { pair, amount0, amount1, rangeWidth = 2 } = req.body;
        // Calculate range based on recent history
        // Fetch last 100 candles
        const candles = await kucoin.fetchOHLCV(pair, '1h', 100); // Using KuCoin data for calc
        const prices = candles.map((c: any) => c.close);
        const { lower, upper } = liquidityManager.calculateOptimalRange(prices, rangeWidth);

        const tx = await uniswap.addLiquidity(pair, amount0, amount1, lower, upper);
        res.json({ success: true, txHash: tx, range: { lower, upper } });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

app.post('/api/defi/liquidity/remove', async (req, res) => {
    try {
        const { tokenId, percent } = req.body;
        const tx = await uniswap.removeLiquidity(Number(tokenId), Number(percent));
        res.json({ success: true, txHash: tx });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

app.post('/api/strategy/backtest', async (req, res) => {
    try {
        console.log("[API] Running Full Backtest...", req.body);
        const symbol = req.body.symbol || "BTC";
        const strategyName = req.body.strategy || "SMAStrategy";
        const timeframe = req.body.timeframe || "1h";

        // Default to last 30 days if not specified
        const now = Date.now();
        const thirtyDays = 30 * 24 * 60 * 60 * 1000;
        const endDate = req.body.endDate ? new Date(req.body.endDate).getTime() : now;
        const startDate = req.body.startDate ? new Date(req.body.startDate).getTime() : (now - thirtyDays);
        const initialBalance = req.body.initialBalance || 10000;

        let strategy = StrategyFactory.get(strategyName);
        if (!strategy && strategyName === "Cointrade") {
             strategy = StrategyFactory.get("Cointrade (External)");
        }

        if (!strategy) {
             return res.status(400).json({ error: `Strategy ${strategyName} not found` });
        }

        const engine = new BacktestEngine();
        const result = await engine.run({
            strategy: strategyName,
            pair: `${symbol}-USD`,
            startDate,
            endDate,
            initialBalance,
            timeframe
        }, strategy);

        res.json(result);
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message || "Backtest failed" });
    }
});

// --- AI Evolution (DeepThinker) ---
const deepThinker = new DeepThinker();

app.post('/api/ai/train', async (req, res) => {
    try {
        const { symbol = "BTC", timeframe = "1h", lookback = 30 } = req.body;

        console.log(`[API] Training DeepThinker for ${symbol} ${timeframe}...`);

        const now = Date.now();
        const start = now - (lookback * 24 * 60 * 60 * 1000); // Days
        const candles = await HistoricalData.fetch(`${symbol}-USD`, timeframe, start, now);

        if (candles.length < 100) {
            return res.status(400).json({ error: "Insufficient data for training" });
        }

        // Run training in background (or await if simple)
        const history = await deepThinker.train(candles);

        res.json({ success: true, history });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

app.post('/api/ai/predict', async (req, res) => {
    try {
        const { symbol = "BTC", timeframe = "1h" } = req.body;

        // Fetch last 100 candles for context
        const now = Date.now();
        const start = now - (5 * 24 * 60 * 60 * 1000);
        const candles = await HistoricalData.fetch(`${symbol}-USD`, timeframe, start, now);

        const prediction = await deepThinker.predict(candles);
        const lastClose = candles[candles.length - 1].close;
        const direction = prediction > lastClose ? "UP" : "DOWN";

        res.json({
            prediction,
            lastClose,
            direction,
            diff: prediction - lastClose
        });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

// --- Arbitrage ---
const arbScanner = new ArbitrageScanner();

app.get('/api/arbitrage/opportunities', async (req, res) => {
    try {
        const { coins = "BTC,ETH,SOL,MATIC" } = req.query;
        const symbols = (coins as string).split(',');
        const opportunities: any[] = [];

        for (const sym of symbols) {
            const opp = await arbScanner.scan(sym);
            if (opp) opportunities.push(opp);
        }

        res.json({ opportunities });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

// --- Risk & Sentiment Analysis ---
const correlationEngine = new CorrelationMatrix(kucoin); // Use KuCoin for price data
const sentimentScanner = new SentimentScanner();

app.get('/api/sentiment/:symbol', async (req, res) => {
    try {
        const symbol = req.params.symbol;
        const result = await sentimentScanner.scan(symbol);
        res.json(result);
    } catch (e: any) {
        res.status(500).json({ error: e.message });
    }
});

app.get('/api/sentiment/market/fng', async (req, res) => {
    try {
        const result = await sentimentScanner.getMarketFearAndGreed();
        res.json(result);
    } catch (e: any) {
        res.status(500).json({ error: e.message });
    }
});

app.get('/api/risk/correlation', async (req, res) => {
    try {
        const { coins = "BTC,ETH,SOL,MATIC,DOGE" } = req.query;
        const symbolList = (coins as string).split(',');

        const data = await correlationEngine.generateMatrix(symbolList);
        res.json(data);
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

const whaleWatcher = new WhaleWatcher();

app.get('/api/risk/whales', async (req, res) => {
    try {
        const events = await whaleWatcher.scanRecentWhaleActivity();
        res.json({ events });
    } catch (e: any) {
        console.error(e);
        res.status(500).json({ error: e.message });
    }
});

// System Status & Submodule Dashboard
app.get('/api/system/status', (req, res) => {
    // Read actual version from VERSION.md if possible, using hardcoded fallback for now
    const currentVersion = "3.5.0";

    res.json({
        version: currentVersion,
        modules: {
            trader: { status: "active", version: "3.0.0", description: "Core Execution Engine" },
            thinker: { status: "active", engine: config.get("trading.active_ai") || "DeepThinker", description: "AI & Signal Generator" },
            analytics: { status: "active", db: "sqlite", description: "Trade Logging & KPIs" },
            notifications: { status: "active", enabled: !!config.get("notifications.enabled"), description: "Discord/Telegram Alerts" },
            risk: { status: "active", version: "1.0.0", description: "Correlation & Sizing" },
            defi: { status: "active", version: "1.2.0", description: "Uniswap V3 LP Manager" }
        },
        exchanges: {
            active: config.get("trading.active_exchange"),
            available: ["robinhood", "kucoin", "binance", "coinbase", "uniswap"]
        },
        submodules: [
            {
                name: "cointrade",
                version: "1.4.2",
                build: "8472a",
                date: "2026-01-10",
                path: "powertrader-ts/backend/src/modules/cointrade",
                description: "Ported technical analysis & strategy components from legacy cointrade repo."
            },
            {
                name: "hyperopt",
                version: "1.1.0",
                build: "192bb",
                date: "2026-02-05",
                path: "powertrader-ts/backend/src/extensions/hyperopt",
                description: "Genetic algorithm for strategy parameter optimization."
            }
        ],
        project_structure: {
            root: "/app/powertrader-ts",
            backend: "/app/powertrader-ts/backend (Node.js API)",
            frontend: "/app/powertrader-ts/frontend (React SPA)",
            data: "/app/powertrader-ts/hub_data (SQLite DBs & Models)"
        }
    });
});

export function startServer() {
    const server = http.createServer(app);
    WebSocketManager.getInstance().initialize(server);

    server.listen(port, () => {
        console.log(`[API] Backend running at http://localhost:${port}`);
    });
}